--- file.json
@@ -1,1 +1,0 @@
-{"User.8d95081c-2752-41de-9677-24e5327f9816": {"id": "8d95081c-2752-41de-9677-24e5327f9816", "created_at": "2024-03-31T20:26:34.788328", "updated_at": "2024-03-31T20:26:34.788452", "__class__": "User"}, "User.4b0522cf-8ba2-4c66-a869-f03839508507": {"id": "4b0522cf-8ba2-4c66-a869-f03839508507", "created_at": "2024-03-31T20:26:37.232708", "updated_at": "2024-03-31T20:26:37.232821", "__class__": "User"}, "User.a1f64cb3-0789-43bb-921a-a4f37e578446": {"id": "a1f64cb3-0789-43bb-921a-a4f37e578446", "created_at": "2024-03-31T20:26:38.636992", "updated_at": "2024-03-31T20:26:38.637144", "__class__": "User"}, "State.de8872db-0bd9-4cbf-8048-181dbbe3fb4d": {"id": "de8872db-0bd9-4cbf-8048-181dbbe3fb4d", "created_at": "2024-03-31T20:27:05.498189", "updated_at": "2024-03-31T20:27:05.498351", "__class__": "State"}, "State.07245f45-fdf6-4403-a5ab-7fe4775c6379": {"id": "07245f45-fdf6-4403-a5ab-7fe4775c6379", "created_at": "2024-03-31T20:27:06.721221", "updated_at": "2024-03-31T20:27:06.721337", "__class__": "State"}, "State.03952379-07da-49ed-9fc4-cf8902fc9b23": {"id": "03952379-07da-49ed-9fc4-cf8902fc9b23", "created_at": "2024-03-31T20:27:07.620426", "updated_at": "2024-03-31T20:27:07.620542", "__class__": "State"}}
