+++ tests/test_models/test_engine/test_file_storage.py
@@ -1,111 +1,141 @@
 #!/usr/bin/python3
-"""test for file storage"""
-import unittest
-import pycodestyle
-import json
-import os
+"""
+Contains the TestFileStorageDocs classes
+"""
+
 from datetime import datetime
 import inspect
 import models
-from models.base_model import BaseModel
-from models.user import User
-from models.state import State
-from models.city import City
+from models.engine import file_storage
 from models.amenity import Amenity
+from models.base_model import BaseModel
+from models.city import City
 from models.place import Place
 from models.review import Review
-from models.engine.file_storage import FileStorage
+from models.state import State
+from models.user import User
+import json
+from os import getenv
+import pep8
+import unittest
+FileStorage = file_storage.FileStorage
+classes = {"Amenity": Amenity, "BaseModel": BaseModel, "City": City,
+           "Place": Place, "Review": Review, "State": State, "User": User}
+
+
+class TestFileStorageDocs(unittest.TestCase):
+    """Tests to check the documentation and style of FileStorage class"""
+    @classmethod
+    def setUpClass(cls):
+        """Set up for the doc tests"""
+        cls.fs_f = inspect.getmembers(FileStorage, inspect.isfunction)
+
+    def test_pep8_conformance_file_storage(self):
+        """Test that models/engine/file_storage.py conforms to PEP8."""
+        pep8s = pep8.StyleGuide(quiet=True)
+        result = pep8s.check_files(['models/engine/file_storage.py'])
+        self.assertEqual(result.total_errors, 0,
+                         "Found code style errors (and warnings).")
+
+    def test_pep8_conformance_test_file_storage(self):
+        """Test tests/test_models/test_file_storage.py conforms to PEP8."""
+        pep8s = pep8.StyleGuide(quiet=True)
+        result = pep8s.check_files(['tests/test_models/test_engine/\
+test_file_storage.py'])
+        self.assertEqual(result.total_errors, 0,
+                         "Found code style errors (and warnings).")
+
+    def test_file_storage_module_docstring(self):
+        """Test for the file_storage.py module docstring"""
+        self.assertIsNot(file_storage.__doc__, None,
+                         "file_storage.py needs a docstring")
+        self.assertTrue(len(file_storage.__doc__) >= 1,
+                        "file_storage.py needs a docstring")
+
+    def test_file_storage_class_docstring(self):
+        """Test for the FileStorage class docstring"""
+        self.assertIsNot(FileStorage.__doc__, None,
+                         "FileStorage class needs a docstring")
+        self.assertTrue(len(FileStorage.__doc__) >= 1,
+                        "FileStorage class needs a docstring")
+
+    def test_fs_func_docstrings(self):
+        """Test for the presence of docstrings in FileStorage methods"""
+        for func in self.fs_f:
+            self.assertIsNot(func[1].__doc__, None,
+                             "{:s} method needs a docstring".format(func[0]))
+            self.assertTrue(len(func[1].__doc__) >= 1,
+                            "{:s} method needs a docstring".format(func[0]))
 
 
 class TestFileStorage(unittest.TestCase):
-    '''This class tests the FileStorage class'''
+    """Test the FileStorage class"""
+    @unittest.skipIf(getenv("HBNB_TYPE_STORAGE") == 'db', "not testing fs")
+    def test_all_returns_dict(self):
+        """Test that all returns the FileStorage.__objects attr"""
+        storage = FileStorage()
+        new_dict = storage.all()
+        self.assertEqual(type(new_dict), dict)
+        self.assertIs(new_dict, storage._FileStorage__objects)
 
-    @classmethod
-    def setUpClass(cls):
-        """Set up for the test class"""
-        cls.storage = FileStorage()
-
-    def setUp(self):
-        """Set up for each test case"""
-        self.storage.reload()
-
-    def tearDown(self):
-        """Tear down after each test case"""
-        try:
-            os.remove("file.json")
-        except FileNotFoundError:
-            pass
-
-    def test_pycodestyle_FileStorage(self):
-        """Test for PEP 8 style"""
-        style = pycodestyle.StyleGuide(quiet=True)
-        result = style.check_files(['models/engine/file_storage.py'])
-        self.assertEqual(result.total_errors, 0, "Fix PEP 8 style issues")
-
-    def test_all(self):
-        """Test the all method"""
-        obj = self.storage.all()
-        self.assertIsNotNone(obj)
-        self.assertEqual(type(obj), dict)
-        self.assertIs(obj, self.storage._FileStorage__objects)
-
+    @unittest.skipIf(getenv("HBNB_TYPE_STORAGE") == 'db', "not testing fs")
     def test_new(self):
-        """Test the new method"""
-        user = User()
-        user.id = 123455
-        user.name = "Kevin"
-        self.storage.new(user)
-        key = f"{user.__class__.__name__}.{str(user.id)}"
-        self.assertIn(key, self.storage._FileStorage__objects)
+        """test that new adds an object to the FileStorage.__objects attr"""
+        storage = FileStorage()
+        save = FileStorage._FileStorage__objects
+        FileStorage._FileStorage__objects = {}
+        test_dict = {}
+        for key, value in classes.items():
+            with self.subTest(key=key, value=value):
+                instance = value()
+                instance_key = instance.__class__.__name__ + "." + instance.id
+                storage.new(instance)
+                test_dict[instance_key] = instance
+                self.assertEqual(test_dict, storage._FileStorage__objects)
+        FileStorage._FileStorage__objects = save
 
-    def test_reload(self):
-        """Test the reload method"""
-        self.storage.save()
-        with open("file.json", 'r') as f:
-            lines = f.readlines()
-        self.storage.save()
-        with open("file.json", 'r') as f:
-            lines2 = f.readlines()
-        self.assertEqual(lines, lines2)
+    @unittest.skipIf(getenv("HBNB_TYPE_STORAGE") == 'db', "not testing fs")
+    def test_save(self):
+        """Test that save properly saves objects to file.json"""
+        storage = FileStorage()
+        new_dict = {}
+        for key, value in classes.items():
+            instance = value()
+            instance_key = instance.__class__.__name__ + "." + instance.id
+            new_dict[instance_key] = instance
+        save = FileStorage._FileStorage__objects
+        FileStorage._FileStorage__objects = new_dict
+        storage.save()
+        FileStorage._FileStorage__objects = save
+        for key, value in new_dict.items():
+            new_dict[key] = value.to_dict()
+        string = json.dumps(new_dict)
+        with open("file.json", "r") as f:
+            js = f.read()
+        self.assertEqual(json.loads(string), json.loads(js))
 
-    def test_get_by_class_and_id(self):
-        """Test the get method by class and ID"""
-        user = User()
-        self.storage.new(user)
-        self.storage.save()
-        retrieved_user = self.storage.get(User, user.id)
-        self.assertEqual(retrieved_user, user)
+    @unittest.skipIf(getenv("HBNB_TYPE_STORAGE") == 'db', "not testing fs")
+    def test_get(self):
+        """ Tests method for obtaining an instance file storage"""
+        storage = FileStorage()
+        dic = {"name": "Vecindad"}
+        instance = State(**dic)
+        storage.new(instance)
+        storage.save()
+        storage = FileStorage()
+        get_instance = storage.get(State, instance.id)
+        self.assertEqual(get_instance, instance)
 
-    def test_get_non_existing_object(self):
-        """Test the get method for non-existing object"""
-        retrieved_obj = self.storage.get(User, "non_existing_id")
-        self.assertIsNone(retrieved_obj)
-
-    def test_get_with_invalid_parameters(self):
-        """Test the get method with invalid parameters"""
-        retrieved_obj = self.storage.get(None, None)
-        self.assertIsNone(retrieved_obj)
-
-    def test_count_all_objects(self):
-        """Test the count method for all objects"""
-        obj1 = BaseModel()
-        obj2 = State()
-        obj3 = City()
-        self.storage.new(obj1)
-        self.storage.new(obj2)
-        self.storage.new(obj3)
-        self.assertEqual(self.storage.count(), self.storage.count())
-
-    def test_count_objects_by_class(self):
-        """Test the count method for objects by class"""
-        obj1 = BaseModel()
-        obj2 = State()
-        obj3 = State()
-        self.storage.new(obj1)
-        self.storage.new(obj2)
-        self.storage.new(obj3)
-        self.assertEqual(self.storage.count(State), self.storage.count(State))
-
-
-if __name__ == "__main__":
-    unittest.main()
+    @unittest.skipIf(getenv("HBNB_TYPE_STORAGE") == 'db', "not testing fs")
+    def test_count(self):
+        """ Tests count method file storage """
+        storage = FileStorage()
+        dic = {"name": "Vecindad"}
+        state = State(**dic)
+        storage.new(state)
+        dic = {"name": "Mexico"}
+        city = City(**dic)
+        storage.new(city)
+        storage.save()
+        c = storage.count()
+        self.assertEqual(len(storage.all()), c)
